requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
ultralytics==8.3.158
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.4
torch==2.2.0
torchvision==0.17.0
python-dotenv=1.1.1